Moved Htaccess management to separte class
