+++ watermark.php
@@ -28,6 +28,7 @@ if (!defined('_PS_VERSION_')) {
 }
 
 use PrestaShop\Module\Watermark\Addons\CategoryFetcher;
+use PrestaShop\Module\Watermark\Htaccess\HtaccessManager;
 
 /**
  * Class Watermark
@@ -50,6 +51,9 @@ class Watermark extends Module
     private $imageTypes = [];
     /** @var array $extensionCache */
     protected $extensionCache = [];
+
+    /** @var HtaccessManager */
+    private $htaccessManager;
 
     /**
      * Watermark constructor.
@@ -103,6 +107,8 @@ class Watermark extends Module
         if (!isset($this->transparency, $this->xAlign, $this->yAlign)) {
             $this->warning = $this->trans('Watermark image must be uploaded for this module to work properly.', [], 'Modules.Watermark.Admin');
         }
+        $this->htaccessManager = new HtaccessManager($this->context->getTranslator());
+
     }
 
     /**
@@ -116,7 +122,7 @@ class Watermark extends Module
             return false;
         }
 
-        $this->writeHtaccessSection();
+        $this->htaccessManager->writeHtaccessSection();
         if (!parent::install() || !$this->registerHook('actionWatermark')) {
             return false;
         }
@@ -132,7 +138,7 @@ class Watermark extends Module
      */
     public function uninstall()
     {
-        if (!$this->removeHtaccessSection()) {
+        if (!$this->htaccessManager->removeHtaccessSection()) {
             $this->context->controller->errors[] = $this->trans('Unable to remove watermark section from .htaccess file', [], 'Modules.Watermark.Admin');
         }
 
@@ -252,66 +258,6 @@ class Watermark extends Module
     }
 
     /**
-     * @return bool|string
-     */
-    public function getAdminDir()
-    {
-        $adminDir = str_replace('\\', '/', _PS_ADMIN_DIR_);
-        $adminDir = explode('/', $adminDir);
-        $len = count($adminDir);
-
-        return $len > 1 ? $adminDir[$len - 1] : _PS_ADMIN_DIR_;
-    }
-
-    /**
-     * @return bool
-     */
-    public function removeHtaccessSection()
-    {
-        $key1 = "\n# start ~ module watermark section";
-        $key2 = "# end ~ module watermark section\n";
-        $path = _PS_ROOT_DIR_ . '/.htaccess';
-        if (file_exists($path) && is_writable($path)) {
-            $s = Tools::file_get_contents($path);
-            $p1 = strpos($s, $key1);
-            $p2 = strpos($s, $key2, $p1);
-            if ($p1 === false || $p2 === false) {
-                return false;
-            }
-            $s = Tools::substr($s, 0, $p1) . Tools::substr($s, $p2 + Tools::strlen($key2));
-            file_put_contents($path, $s);
-        }
-
-        return true;
-    }
-
-    /**
-     * Write the watermark section to the .htaccess file
-     *
-     * @return boolean
-     */
-    public function writeHtaccessSection()
-    {
-        $adminDir = $this->getAdminDir();
-        $source = "\n# start ~ module watermark section
-<IfModule mod_rewrite.c>
-RewriteEngine On
-RewriteCond expr \"! %{HTTP_REFERER} -strmatch '*://%{HTTP_HOST}*/$adminDir/*'\"
-RewriteRule [0-9/]+/[0-9]+\\.jpg$ - [F]
-</IfModule>
-# end ~ module watermark section\n";
-
-        $path = _PS_ROOT_DIR_ . '/.htaccess';
-        if (false === file_put_contents($path, $source . file_get_contents($path))) {
-            $this->context->controller->errors[] = $this->trans('Unable to add watermark section to the .htaccess file', [], 'Modules.Watermark.Admin');
-
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
      * @return string
      *
      * @throws PrestaShopDatabaseException
@@ -322,8 +268,8 @@ RewriteRule [0-9/]+/[0-9]+\\.jpg$ - [F]
     public function getContent()
     {
         // Modify htaccess to prevent download of original pictures
-        $this->removeHtaccessSection();
-        $this->writeHtaccessSection();
+        $this->htaccessManager->removeHtaccessSection();
+        $this->htaccessManager->writeHtaccessSection();
 
         $html = '';
         if (Tools::isSubmit('btnSubmit')) {
